Broke down test cases for ServiceControlClientImpl into three pieces by feature
